Fix unit test and integration test errors
