+++ tests/integration_test/ha_test_cases.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+# TODO:: remove this file
 
 # HA test cases (referenced from NVFlare 2.1 Test Plan- High Availability)
